requirements.txt
pyyaml==6.0
# torch==1.11.0
# torchvision==0.12.0
# pandas==1.4.2

termcolor==1.1.0
torchmetrics==0.9.3

Pillow==8.2.0
imageio==2.9.0

numpy==1.20.3
scikit-image==0.18.3
scipy==1.6.3
seaborn==0.11.2
six==1.15.0
torch==1.9.0
torchvision==0.10.0
pandas==1.3.2
opencv-python==4.5.3.56
tqdm

notebook

scikit-image # For image io and transforms
pandas # For easier csv parsing